+++ src/extension.ts
@@ -1,5 +1,6 @@
 import * as vscode from "vscode";
 import { MezDefinitionProvider } from "./definitionProvider";
+import { MezDocumentSymbolProvider } from "./documentSymbolProvider";
 import { MezReferenceProvider } from "./referenceProvider";
 import { MezSymbolIndex } from "./symbolIndex";
 
@@ -13,7 +14,11 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
   context.subscriptions.push(
     index,
     vscode.languages.registerDefinitionProvider(selector, new MezDefinitionProvider(index)),
-    vscode.languages.registerReferenceProvider(selector, new MezReferenceProvider(index))
+    vscode.languages.registerReferenceProvider(selector, new MezReferenceProvider(index)),
+    vscode.languages.registerDocumentSymbolProvider(
+      selector,
+      new MezDocumentSymbolProvider(index)
+    )
   );
 }